+++ src/ca/mcgill/ecse420/a1/DiningPhilosophersNoStarvation.java
@@ -94,7 +94,7 @@ public class DiningPhilosophersNoStarvation {
         }
       }
       // To see # of times each philosopher ate, comment out the other System.out.println() lines
-      System.out.println("Philosopher " + id + " ate " + ate + " times and waited for " + (double)totalWait/1000000000.0 + " seconds");
+      System.out.println("Philosopher " + id + " ate " + ate + " times and waited for " + totalWait/1000000000.0 + " seconds");
     }
   }
 }
